Move screenshot callback to own file
