+++ src/cli.ts
@@ -4,10 +4,11 @@ import chalk from "chalk";
 import clear from "clear";
 import program from "commander";
 import figlet from "figlet";
-import fs from "fs-extra";
 import { BrowserRunner } from "./BrowserRunner";
+import { CONFIG } from "./config";
 import { Server } from "./io/Server";
-import { planJob } from "./planner";
+import { buildScreenshotCallback } from "./screenshot";
+import { Job, BrowserStep } from "./types";
 
 clear();
 
@@ -18,16 +19,50 @@ console.log(
 program.version("0.0.1").description("Effortless smoke tests");
 
 program
-  .command("run <source> [destination]")
+  .command("run")
   .description("run a test")
   .action(async (source, destination) => {
-    const events = JSON.parse(await fs.readFile(source, "utf8"));
-    const job = planJob(events);
+    const steps: BrowserStep[] = [
+      {
+        selector: {
+          xpath: '//*[@id="username"]'
+        },
+        type: "type",
+        value: "tomsmith"
+      },
+      {
+        selector: {
+          xpath: '//*[@id="password"]'
+        },
+        type: "type",
+        value: "SuperSecretPassword!"
+      },
+      {
+        selector: {
+          xpath: '//*[@id="login"]/button'
+        },
+        type: "click"
+      }
+    ];
 
+    const job: Job = {
+      href: `${CONFIG.testUrl}/login`,
+      steps
+    };
     const server = new Server();
-    const runner = new BrowserRunner({ server });
+    await server.listen();
 
+    const takeScreenshot = buildScreenshotCallback(1000);
+
+    const callbacks = {
+      beforeStep: [takeScreenshot],
+      afterRun: [takeScreenshot]
+    };
+
+    const runner = new BrowserRunner({ callbacks, server });
     await runner.run(job);
+
+    await runner.close();
   });
 
 program.parse(process.argv);
